+++ src/main/java/wooteco/subway/station/service/StationService.java
@@ -1,8 +1,8 @@
 package wooteco.subway.station.service;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.station.dao.StationDao;
-import wooteco.subway.station.dao.dto.StationDto;
+import wooteco.subway.station.repository.StationRepository;
+import wooteco.subway.station.repository.dto.StationDto;
 import wooteco.subway.station.domain.Station;
 
 import java.util.List;
@@ -11,25 +11,25 @@ import java.util.stream.Collectors;
 @Service
 public class StationService {
 
-    private StationDao stationDao;
+    private StationRepository stationRepository;
 
-    public StationService(StationDao stationDao) {
-        this.stationDao = stationDao;
+    public StationService(StationRepository stationRepository) {
+        this.stationRepository = stationRepository;
     }
 
     public StationDto save(StationDto stationDto) {
-        Station station = stationDao.save(new Station(stationDto.getName()));
+        Station station = stationRepository.save(new Station(stationDto.getName()));
         return StationDto.from(station);
     }
 
     public List<StationDto> findAll() {
-        List<Station> stations = stationDao.findAll();
+        List<Station> stations = stationRepository.findAll();
         return stations.stream()
                 .map(station -> StationDto.from(station))
                 .collect(Collectors.toList());
     }
 
     public void delete(Long id) {
-        stationDao.delete(id);
+        stationRepository.delete(id);
     }
 }
